test(plugin-dashboard): move the #4426 type pins into src, not a narrow typetests project
CI caught the first attempt: `scripts/__tests__/check-type-check-coverage.test.ts`
pins objectui#4291's ratchet as a repository-state test — a
`tsconfig.typetests.json` reappearing ANYWHERE turns it red. The gate script
permits the shape; the ratchet forbids a new user. The narrow project is removed
and `type-check` goes back to plain `tsc --noEmit`.

The compile-time assertions move to `src/domPassthroughPins.ts`, which the
package's own `tsc --noEmit` already compiles — `widgets/toDomProps.ts`'s shape
in @object-ui/fields. They are `type`-only, so they emit zero runtime bytes.
Reverting either `extends` turns them red under the ordinary type-check.

The test file keeps only its RUNTIME assertions; its `@ts-expect-error` cases are
dropped rather than left in a file nothing compiles, which would have read as
coverage while checking nothing (objectui#3181).
